[MIG] website_sale_loyalty_suggestion_wizard: Migration to version 18.0
TT54533
