+++ website_sale_loyalty_suggestion_wizard/README.rst
@@ -99,13 +99,13 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-.. |maintainer-chienandalu| image:: https://github.com/chienandalu.png?size=40px
-    :target: https://github.com/chienandalu
-    :alt: chienandalu
+.. |maintainer-pilarvargas-tecnativa| image:: https://github.com/pilarvargas-tecnativa.png?size=40px
+    :target: https://github.com/pilarvargas-tecnativa
+    :alt: pilarvargas-tecnativa
 
 Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
 
-|maintainer-chienandalu| 
+|maintainer-pilarvargas-tecnativa| 
 
 This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/18.0/website_sale_loyalty_suggestion_wizard>`_ project on GitHub.
