+++ website_sale_loyalty_suggestion_wizard/templates/promotion_templates.xml
@@ -21,7 +21,7 @@
         inherit_id="website_sale.cart_lines"
         name="Configure Promotion"
     >
-        <xpath expr="//table[@id='cart_products']" position="before">
+        <xpath expr="//div[@id='cart_products']" position="before">
             <t t-if="promotion_id">
                 <div
                     class="alert alert-warning alert-dismissible fade show"
@@ -119,17 +119,20 @@
         customize_show="True"
         name="Promotions suggestions"
     >
-        <xpath expr="//th[hasclass('td-img')]" position="before">
+        <xpath expr="//t[@t-if='line.product_id']" position="before">
             <t
                 t-set="promotions"
                 t-value="website_sale_order.website_order_line.suggested_promotion_ids.filtered(lambda x: x.is_published)[:1].id"
             />
-            <t t-if="promotions">
+            <!-- <t t-if="promotions">
                 <th />
-            </t>
+            </t> -->
         </xpath>
-        <xpath expr="//td[hasclass('td-img')]" position="before">
-            <td class="td-promo-suggestion" t-if="promotions">
+        <xpath expr="//t[@t-if='line.product_id']/div" position="before">
+            <div
+                class="td-promo-suggestion d-flex align-items-center"
+                t-if="promotions"
+            >
                 <t
                     t-set="promo"
                     t-value="line.suggested_promotion_ids.filtered(lambda x: x.is_published)[:1].id"
@@ -143,9 +146,9 @@
                         class="fa fa-gift fa-2x mx-auto"
                     />
                 </t>
-            </td>
+            </div>
         </xpath>
-        <xpath expr="//table[@id='cart_products']" position="after">
+        <xpath expr="//div[@id='cart_products']" position="after">
             <t
                 t-set="suggested_promos"
                 t-value="website_sale_order.website_order_line.suggested_promotion_ids.filtered('is_published')"
